refactor(modules): move auth related components into auth module
