+++ src/account/account.module.ts
@@ -1,7 +1,4 @@
 import { Module } from '@nestjs/common';
-import { VerifyLoginService } from './services/verify-login.service';
 
-@Module({
-    providers: [VerifyLoginService],
-})
+@Module({})
 export class AccountModule {}
