--- src/account/interfaces/auth.interface.ts
@@ -1,5 +1,0 @@
-export interface LoginRequest {
-    wallet: string;
-    timestamp: number;
-    signature: string;
-}
