+++ src/auth/auth.interface.ts
@@ -1,0 +1,13 @@
+export interface JwtPayload {
+    wallet: string;
+}
+
+export interface AuthenticatedUser {
+    wallet: string;
+}
+
+export interface LoginRequest {
+    wallet: string;
+    timestamp: number;
+    signature: string;
+}
